bugfix: fix from reviwe
