+++ src/console/index.ts
@@ -2,3 +2,4 @@ export * from './messages/massages';
 export * from './helpers/helpers';
 export * from './styles/styles';
 export * from './command/command';
+export * from './utils/commandHandlers';
